// src/config/config.types.ts
export type EnvironmentType = 'development' | 'production' | 'test' | 'staging';